+++ wakamiti-engine/wakamiti-core/src/main/java/es/iti/wakamiti/core/datatypes/assertion/BinaryNumberAssertProvider.java
@@ -145,7 +145,7 @@ public class BinaryNumberAssertProvider<T extends Comparable<T>> extends Abstrac
      * {@inheritDoc}
      */
     @Override
-    protected LinkedList<String> regex(Locale locale) {
+    public LinkedList<String> regex(Locale locale) {
         return Arrays.stream(expressions())
                 .map(exp -> ExpressionMatcher.computeRegularExpression(bundle(locale).getString(exp)))
                 .map(exp -> exp.replace(VALUE_WILDCARD, numberRegexProvider.apply(locale)))