+++ wakamiti-engine/wakamiti-core/src/main/java/module-info.java
@@ -8,6 +8,7 @@ import es.iti.wakamiti.api.extensions.*;
 import es.iti.wakamiti.core.DefaultWakamitiAPI;
 import es.iti.wakamiti.core.datatypes.WakamitiCoreTypes;
 import es.iti.wakamiti.core.datatypes.assertion.WakamitiAssertTypes;
+import es.iti.wakamiti.core.datatypes.duration.WakamitiDurationType;
 import es.iti.wakamiti.core.gherkin.GherkinPlanBuilder;
 import es.iti.wakamiti.core.gherkin.GherkinRedefinitionPlanTransformer;
 import es.iti.wakamiti.core.gherkin.GherkinResourceType;
@@ -44,7 +45,7 @@ open module es.iti.wakamiti.core {
     provides ResourceType with GherkinResourceType;
     provides PlanBuilder with GherkinPlanBuilder;
     provides PlanTransformer with GherkinRedefinitionPlanTransformer;
-    provides DataTypeContributor with WakamitiCoreTypes, WakamitiAssertTypes;
+    provides DataTypeContributor with WakamitiCoreTypes, WakamitiAssertTypes, WakamitiDurationType;
     provides PropertyEvaluator with GlobalPropertyEvaluator, StepPropertyEvaluator;
 
 }